static_analysis/requirements.txt
psycopg2-binary>=2.9.0
pandas>=1.5.0
matplotlib>=3.5.0
python-dotenv>=0.19.0
folium>=0.14.0
geopandas>=0.13.0
shapely>=2.0.0
seaborn>=0.12.0


